requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
